Refactored codebase to use OOP paradigm
I refactored the code to improve its organization while keeping the core 
functionality of the game intact.

I created a Score class to track the scores of both the player and the computer. 
Previously, these scores were managed in the global scope, which is not 
considered good practice. By storing the scores in an object, we remove them 
from the global scope and encapsulate them with methods responsible for 
checking and manipulating them.

Additionally, I introduced a Round class to encapsulate the data and methods
related to an individual round. The class includes data members such as the 
player and computer moves, the available moves for the computer, and the 
results of each round.

Furthermore, I created a UI class to encapsulate references to various HTML 
IDs and implemented methods that utilize these IDs to display the scores, 
round results, and match results to the player.

Overall, by implementing the object-oriented programming (OOP) paradigm, 
the code becomes more readable and easier to maintain.